Created an AppleBasket Game Object
In trying to come up with a way to "mash-up" this project with the ones we did previously, I wasn't able to come up with something regarding Apple Picker.  So, what I did was make some references to that project, with an apple basket that's present at all times, and a tree that's shown in the Castle_1 Prefab.